# documentation-swagger/blog-api-swagger.yaml
swagger: '2.0'
info:
  description: 'Documentação de uma API para blogs. Este projeto está em desenvolvimento, qualquer problema ou dúvida entre em contato.'
  version: '1.0'
  title: Blog API
  contact:
    name: Jovani Brasil
    url: http://jovanibrasil.com
    email: jovanibrasil@gmail.com
host: 'localhost:8081'
basePath: /
...
paths:
...
  /users:
    post:
      tags:
        - user-controller
      summary: Cria um usuário.
      operationId: saveUserUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Header para JWT
          required: false
          type: string
        - in: body
          name: userForm
          description: userForm
          required: true
          schema:
            $ref: '#/definitions/UserForm'
      responses:
        '200':
          description: Usuário criado com sucesso.
          schema:
            type: object
  '/users/{userName}':
    get:
      tags:
        - user-controller
      summary: Busca usuário por nome.
      operationId: getUserUsingGET
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Header para JWT
          required: false
          type: string
        - name: userName
          in: path
          description: userName
          required: true
          type: string
      responses:
        '200':
          description: Usuario encontrado e retornado.
          schema:
            $ref: '#/definitions/UserDTO'
        '404':
          description: Usuário não encontrado.
    delete:
      tags:
        - user-controller
      summary: Remove um usuário.
      operationId: deleteUserUsingDELETE
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: Authorization
          in: header
          description: Header para JWT
          required: false
          type: string
        - name: userName
          in: path
          description: userName
          required: true
          type: string
      responses:
        '204':
          description: Usuário removido.
        '404':
          description: Usuário não encontrado.
definitions:
...
  UserDTO:
    type: object
    properties:
      email:
        type: string
      fullUserName:
        type: string
      githubUserName:
        type: string
      googlescholarLink:
        type: string
      linkedinUserName:
        type: string
      phone:
        type: string
      userId:
        type: integer
        format: int64
      userName:
        type: string
  UserForm:
    type: object
    properties:
      email:
        type: string
      fullUserName:
        type: string
      githubUserName:
        type: string
      googlescholarLink:
        type: string
      linkedinUserName:
        type: string
      phone:
        type: string
      userId:
        type: integer
        format: int64
      userName:
        type: string